+++ lib/src/db/gas/notices/agt_nominations.dart
@@ -1,0 +1,6 @@
+
+
+// ignore_for_file: dangling_library_doc_comments
+
+/// Go to https://infopost.enbridge.com/InfoPost/AGHome.asp?Pipe=AG
+/// Historical Downloads > Capacity > Operationally Available